feat: style 정리 완료
