+++ src/components/index.ts
@@ -1,1 +1,1 @@
-export {default as Layout} from './layout'
+export { default as Layout } from "./layout";
