+++ src/components/layout/index.styles.ts
@@ -6,8 +6,5 @@ export const wrapper = css`
   max-width: 600px;
   min-width: 320px;
   height: 100vh; /* 화면 전체 높이를 차지 */
-  display: flex; /* 가운데 정렬을 위해 flex 사용 */
-  align-items: center; /* 세로 중앙 정렬 */
-  justify-content: center; /* 가로 중앙 정렬 */
   background-color: white; /* 회색 배경 */
 `;
