+++ src/pages/index.ts
@@ -1,1 +1,1 @@
-export {default as Login} from './login'
+export { default as Login } from "./login";
